backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
scikit-learn>=1.5.0
pandas>=2.0.3
numpy>=1.24.4
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.23
textblob>=0.17.1
joblib>=1.3.2
python-multipart>=0.0.6
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
tqdm>=4.65.0
pymysql>=1.1.0

